Refactor seq2seq logic and tests into separate files
